addressing code review comments. reimplemented relative texture placement, removed unnecessary using statements, moved MaxGLTFMaterialExporter implementation to separate file, used safe export parameter parsing, changed TextureUtilities to be a static class, added static PathUtilities implementation
