+++ 3ds Max/Max2Babylon/Tools/Tools.cs
@@ -1,5 +1,4 @@
 ﻿using Autodesk.Max;
-using BabylonExport.Entities;
 using Max2Babylon.Extensions;
 using SharpDX;
 using System;
@@ -8,7 +7,6 @@ using System.IO;
 using System.Linq;
 using System.Reflection;
 using System.Text;
-using System.Text.RegularExpressions;
 using System.Windows.Forms;
 
 namespace Max2Babylon
@@ -1022,44 +1020,6 @@ namespace Max2Babylon
 
         #region File Path
 
-        /// <summary>
-        /// Creates a relative path from one file or folder to another. Input paths that are directories should have a trailing slash.
-        /// </summary>
-        /// <param name="fromPath">Contains the directory that defines the start of the relative path. Directories should have a trailing slash.</param>
-        /// <param name="toPath">Contains the path that defines the endpoint of the relative path. Directories should have a trailing slash.</param>
-        /// <returns>The relative path from the start directory to the end path.</returns>
-        /// <exception cref="ArgumentNullException"><paramref name="toPath"/> is <c>null</c>.</exception>
-        /// <exception cref="UriFormatException"></exception>
-        /// <exception cref="InvalidOperationException"></exception>
-        public static string GetRelativePath(string fromPath, string toPath)
-        {
-            if (string.IsNullOrEmpty(fromPath))
-                return toPath;
-            if (string.IsNullOrEmpty(toPath))
-                throw new ArgumentNullException("toPath");
-
-            Uri fromUri = new Uri(fromPath);
-            Uri toUri = new Uri(toPath);
-
-            if (fromUri.Scheme != toUri.Scheme)
-                return toPath;
-
-            Uri relativeUri = fromUri.MakeRelativeUri(toUri);
-            string relativePath = Uri.UnescapeDataString(relativeUri.ToString());
-
-            if (string.Equals(toUri.Scheme, Uri.UriSchemeFile, StringComparison.OrdinalIgnoreCase))
-                relativePath = relativePath.Replace(Path.AltDirectorySeparatorChar, Path.DirectorySeparatorChar);
-
-            return relativePath;
-        }
-
-
-        public static string UnformatPath(string formattedPath)
-        {
-            string newPath = formattedPath;
-            newPath = Regex.Replace(formattedPath, @"[()]", string.Empty);
-            return newPath;
-        }
 
         public static string FormatPath(string absolutePath)
         {
@@ -1116,33 +1076,6 @@ namespace Max2Babylon
 
             return string.Format(@"({0}){1}", dirName, path);
         }
-
-        public static bool IsBelowModelPath(string folderPath,string modelPath)
-        {
-            string modelFolderPath = Path.GetDirectoryName(modelPath);
-            if (folderPath.StartsWith(modelFolderPath))
-            {
-                return true;
-            }
-
-            return false;
-        }
-
-        public static string GetPathRelativeToModel(string folderPath, string modelPath)
-        {
-            Uri path1 = new Uri(modelPath);
-
-            if (!folderPath.EndsWith("\\"))
-            {
-                folderPath += "\\";
-            }
-
-            Uri path2 = new Uri(folderPath);
-            Uri diff = path1.MakeRelativeUri(path2);
-            return diff.OriginalString;
-        }
-
-
         #endregion
     }
 }
